LayoutTests/discovery/TC-BasicServerServiceDiscovery-expected.txt

Tests Acceptable UA behavior when Basic Server Service is dicovered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() first call to ok callback.
PASS srvs.servicesAvailable==0 is true
PASS Add dev callback.
PASS navigator.getNetworkServices() Second call to ok callback.
PASS srvs.servicesAvailable>=1 is true
PASS successfullyParsed is true

TEST COMPLETE
